lib: Unify read_unsafe/write_unsafe impls across backends
Move the UniqueMmioPointer and SharedMmioPointer read/write method
implementations from each backend module into lib.rs, eliminating the
triplication. Each backend now only provides its MmioOps
implementation; the pointer methods are defined once and dispatch
through the conditionally re-exported backend.

Functional change: The volatile backend now routes sliced access (for
types that are not 1, 2, 4, or 8 bytes in size, or via read_unsafe() /
write_unsafe()) through read_sliced() / write_sliced() as the other
backends do, instead of relying on Rust's (unspecified) implementation
of that via volatile_read() / volatile_write(). The access pattern
observed by the MMIO device might change, but was not defined in the
first place.
