+++ src/backend/custom.rs
@@ -45,8 +45,6 @@
 //! set_mmio_ops!(MyMmioBackend);
 //! ```
 use crate::backend::mmio_ops::MmioOps;
-use crate::{SharedMmioPointer, UniqueMmioPointer};
-use zerocopy::{FromBytes, Immutable, IntoBytes};
 
 unsafe extern "Rust" {
     fn __safe_mmio_read_u8(src: *const u8) -> u8;
@@ -128,7 +126,7 @@ macro_rules! set_mmio_ops {
 }
 
 /// MmioOps backend delegating to consumer-provided extern functions.
-struct Ops;
+pub struct Ops;
 
 // SAFETY: Each method delegates to the consumer-provided extern function of the matching width.
 impl MmioOps for Ops {
@@ -178,57 +176,6 @@ impl MmioOps for Ops {
         unsafe {
             __safe_mmio_write_u64(dst, value);
         }
-    }
-}
-
-impl<T: FromBytes + IntoBytes> UniqueMmioPointer<'_, T> {
-    /// Performs an MMIO read and returns the value.
-    ///
-    /// If `T` is exactly 1, 2, 4 or 8 bytes long then this will be a single operation. Otherwise
-    /// it will be split into several, reading chunks as large as possible.
-    ///
-    /// Note that this takes `&mut self` rather than `&self` because an MMIO read may cause
-    /// side-effects that change the state of the device.
-    ///
-    /// # Safety
-    ///
-    /// This field must be safe to perform an MMIO read from.
-    pub unsafe fn read_unsafe(&mut self) -> T {
-        // SAFETY: self.regs is always a valid and unique pointer to MMIO address space.
-        unsafe { Ops::read(self.regs) }
-    }
-}
-
-impl<T: Immutable + IntoBytes> UniqueMmioPointer<'_, T> {
-    /// Performs an MMIO write of the given value.
-    ///
-    /// If `T` is exactly 1, 2, 4 or 8 bytes long then this will be a single operation. Otherwise
-    /// it will be split into several, writing chunks as large as possible.
-    ///
-    /// # Safety
-    ///
-    /// This field must be safe to perform an MMIO write to.
-    pub unsafe fn write_unsafe(&mut self, value: T) {
-        // SAFETY: self.regs is always a valid and unique pointer to MMIO address space.
-        unsafe {
-            Ops::write(self.regs, value);
-        }
-    }
-}
-
-impl<T: FromBytes + IntoBytes> SharedMmioPointer<'_, T> {
-    /// Performs an MMIO read and returns the value.
-    ///
-    /// If `T` is exactly 1, 2, 4 or 8 bytes long then this will be a single operation. Otherwise
-    /// it will be split into several, reading chunks as large as possible.
-    ///
-    /// # Safety
-    ///
-    /// This field must be safe to perform an MMIO read from, and doing so must not cause any
-    /// side-effects.
-    pub unsafe fn read_unsafe(&self) -> T {
-        // SAFETY: self.regs is always a valid pointer to MMIO address space.
-        unsafe { Ops::read(self.regs) }
     }
 }
