+++ src/backend/mod.rs
@@ -9,3 +9,10 @@ pub mod custom;
 pub mod mmio_ops;
 #[cfg(all(not(target_arch = "aarch64"), not(feature = "custom-mmio")))]
 pub mod volatile;
+
+#[cfg(all(target_arch = "aarch64", not(feature = "custom-mmio")))]
+pub use aarch64::Ops;
+#[cfg(feature = "custom-mmio")]
+pub use custom::Ops;
+#[cfg(all(not(target_arch = "aarch64"), not(feature = "custom-mmio")))]
+pub use volatile::Ops;
